fix(received): 임포트문 수정

+++ src/main/java/coolpotatoes/coolpotatoes_backend/entity/Received.java
@@ -4,10 +4,11 @@ import coolpotatoes.coolpotatoes_backend.constants.AchievementType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.annotation.Id;
+
 
 import java.time.LocalDateTime;
 
@@ -17,7 +18,7 @@ import java.time.LocalDateTime;
 public class Received {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int receivedId;
+    private Long receivedId;
 
     private String certificateName;
